My fix
General recommendation:
Too many white lines
L32 (Constructor): Some stuff are created here but other are created by default.
L40: This could be moved directly as a default value in line 27.
L41: This could be moved directly as a default value in line 28.
L45-66 (Broadcast method): Switch is hard to read here, better us if/else
L48+61+72: Why do you need to always add \n to msg? Also instead of repeating it every time you could just put msg = msg + “\n” at the beginning of the method.
L90: this could be moved directly in line 86.
L132: Similar to L90, move directly to L134 but with ++subordinateCounter and at L119 initialize with 1.
L168-172: Is this even possible?
L174: temp isn’t really helpful for a variable name and it’s used only once.
L185-191 does exactly the same as 177-181: This can be moved after the if/else statement.
L142: This method should handle the error but the catch statement is missing. Also you’re already telling the compiler that the method can throw an exception, so handling the error is not really needed (but I think it would be preferred).
L224: “.toUpperCase” if you’re checking if it’s an empty string it’s not needed.
L231: Why are you printing an empty string here?
L207-238: There are some repetitions of code in both if and else part.
L246-247: This could be done in the same line.
L251: if you need the index, would be better to use a normal for loop then. What happens if msg.equals(“Y”) ? Also it should be if, else if, else if, else to be improve readability.